requirements.txt
asgi_lifespan==1.0.1
beanie==1.27.0
fastapi==0.115.2
httpx==0.27.2
mongomock-motor==0.0.34
motor==3.6.0
passlib==1.7.4
pre-commit==3.8.0
pydantic-settings==2.5.2
PyJWT==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-multipart==0.0.12
starlette==0.39.2
uvicorn==0.31.1
